refactor: replace html-view.ts&css from css-view into html-view folder
