+++ RSS-CSS-Selectors/src/classes/view/main/html-view/html-view.ts
@@ -4,7 +4,7 @@ import {
   IElementCreator,
   IElementCreatorParam,
 } from "../../../util/element-creator/element-creator-types";
-import { NumbersLineView } from "./numbers-line-view/numbers-line-view";
+import { NumbersLineView } from "../css-view/numbers-line-view/numbers-line-view";
 
 const CSS_CLASSES = ["view"];
 const CSS_CLASSES_CSS_TOP_LINE = ["view__top-line"];
